update membership destroy logic to check for moderator or current user's own membership

+++ app/controllers/api/memberships_controller.rb
@@ -23,9 +23,9 @@ class Api::MembershipsController < ApplicationController
   def destroy
     pool = Pool.find_by(id: params[:pool_id])
 
-    # unless pool.moderator_id == current_user.id
-    #   return render json: ['Only moderators are allowed to remove members.'], status: 401
-    # end
+    unless pool.moderator_id == current_user.id || current_user.id == params[:user_id]
+      return render json: ['Only moderators are allowed to remove other members.'], status: 401
+    end
 
     membership = Membership.find_by(pool_id: params[:pool_id], user_id: params[:user_id])
     membership.destroy if membership
